Changeset for converting presto statistics to qpmodel format.
This changeset adds functionality to deserialize a JSON representation
of Presto statistical information for a given table, convert it to the
qpmodel format and add it to the sysstats table in the catalog.

This changeset also contains a function to serialize an entire sysstats
table and write it to a specified file.

The unit test suite has been modified to test the format conversion,
but not the serialization/write-to-disk function.