Changeset with changes for TPC-H 1G Statistical Data (#98)
This changeset contains two files that have the statistics data
for the TPC-H 1G table data set and modifications necessary to
successfully read in this data. It also contains modifications
to the TPC-H unit tests to test reading and using this data.

Finally, this changeset contains the "expected" results files
for the RunFolderAndVerify unit test for TPC-H 1G data.