Add cardinality estimation for IN expression (#104)
Added CE for IN expression. This is done by understanding IN expression as multiple equality expressions connected by OR. Thus, the selectivity is the sum of selectivity for each equality.

Results for TPCH benchmark test changed after this change in CE, namely q12 and q16. The expected results are updated accordingly. All unit tests are able to pass.