force type coerce for number type comparison during Bind()
This is needed because dynamic data types do not support double compares with decimal. A more robust solution is to do the type conversion at run time because that depends the actual value: decimal has better precision and double has better range, if double if out of decimal's range, we shall convert both to double; otherwise they shall be converted to decimals.