Fix join CE with subquery as join children (#122)
Current join CE does not handle leaf node as subquery. This change fixes the issue by pull BaseTableRef from subquery. 

A related problem is different join order may yield different CE for the same memo group, because we decompose a nary-join CE to multiple binary join CE and the binary algorithm is not associative. This actually violates the assumption that group members sharing the same cardinality. This is tracked by #129.