Changeset that adds imdb statistics for jobench suite This changeset contains the statistics for the jobench suite tables. It also makes basic modifications to the Unit Test TestJobench to use the statistics and the expected plans for the jobench queries required for regression testing.
A try/catch structure was added to the method ReadCsvLine.
This structure will catch errors during parse of "table
data" and report the error encountered as well as the file
and record where the error was found.

Finally, a couple of small fixes were made to the JSON post-
processing in Statis.cs for errors that were not caught in
previous unit tests.