Avoid creating SQLStatement twice for ExecSQLList() (#135)
This is for #120

ExecSQLList() pass the statement text to ExecSQL, causing the statement object to be created twice.

The solution is creating an overload of ExecSQL for directly taking in statement object. Adjustment to list execution is also made to avoid the default query option be falsely assigned to ExecSQL even when the option for the outer function is null.