remove copyoutCounter_ and simplify memo validation logics
copyoutCounter_is used to tracking which subquery uses which memo and a side-effect is to tell if the memo has been copy'ed out - so we can skip validation because memo is modified. However, we now do not modify memo in place so we keep memo untouched and can be validated any time. We also use a more straightfoward way to map subquery to its memo.