add histogram construction from min/max values
This is needed to load presto stats which does not contain histogram but only min/max. The histogram can reproduced with even buckets between these two values. Currently there is an issue with presto json stats loading where most data are zero or null - we will hook up this method to stats loading after we fix that.

Along the line, this patch also fix the off by 1 issue of #buckets.