Remove unnecessary redistribution node
Previously join force redistribution on top of table scan for distributed table, but it is not necessary when the key match the distribution.
Now the match is checked, only when the data distribution does not agree with the join keys, a redistribution node is added.

Also, data insertion is modified. Previously, even for distributed table all the rows are added onto the first partition (0), but with distributed table scan all have redistribution node on top, it is still working. Now redistribution may not be present, the original insertion must follow the distributed column.

Additional unittest is added to verify the change.