separate EXISTS subquery and scalar subquery in parser rules
they used to be together as arith_expr but this is problematic as we don't allow EXISTS subquery to do +/- as other arith_expr but scalar subquery is ok to do that.  Now we pull EXISTS subquery to pred_expr and keep scalar query in arith_expr. This also simplify parser logics handling NOT prefix for EXISTS subquery.