New property enforcement framework (#185)
This new version of property enforcement serves as the first step for adding distribution property into memo optimizer.

Previously, for each physical member, there are two costs: cost0 and cost1, which generally corresponds to null property supplied and property supplied. Property can be either directly supplied or supplied through propagation. This can work when only order is considered as property, meaning that a property can be either supplied or not. This binary determination results in the maximum of two different possibilities for each physical member (cost0 and cost1).
However, when distribution is added, property requirement can be partially supplied by a physical member, leading to more than two possible paths. This means the comparison between the inclusive cost of different enforcement nodes will become more complicated.

Given these circumstances, a list of pathways is maintained as an attribute for each member. This keep track of the different supported property and required property pairs of that particular node.
The new method includes enforcement nodes as members and they are created when certain requirement is passed to the group when calculating the min cost member. Hence, the calculation is no longer for each group, but for each group-property pair, and the optimal member for each pair can only be determined after all its child pairs are determined.

Another major change is on how to determine the relationship between physic node and property. The new physic node-property interface is no longer returning the required property, supplied property and propagated property individually, but doing it according to the requested property. If the requirement is satisfied, then the corresponding child requirements are returned.
The property that are less strict is directly derived from the current requirement rather than suggested by possible supplied property of physic node. For each required property, if there is order part, then sorted node is added to the member list as enforcement node, (if there is distribution requirement, for singleton, gather is added, for hash distribution, redistribution is added since redistribution, similar to gather, does not need to know the distribution of the child.) In other word, property is always required, not supplied.