fix race condition of distributed execution
In the real execution, we will have to serialize all information we want the receiver side to deserialize, including physic node an logic node and could be some more. So to emulate this behavior correctly, we shall either having one copy for each participanting thread or make sure their shared copy not change. We do the second way by moving CreateKeyList() to the last step in logic join creation, instead of every physic join doing that separately.