teach all expression in the same use the same card (#195)
Expressions in the same group logically shall have the same cardinality. This is not guaranteed however if we compute CE for both AxBxC and AxCxB - because the join CE formula does guarantee association. To fix this, we compoute the first expression in the group and set all others to this value. This is not perfect though, as the order of expression in the group actually decides the cardinalty.