add GetHashCode() for AggCountStar (#226)
as we override its Equals() function and these two functions shall be paired. It also simplifies Equals() comparison logics.

This change also expose a memo issue which uses certain hash combination as a node's signature. A query in TestJoin() with many similar count(*) confuses it. I have to disable memo for this query and shall fix it in another checkin.