REMOVE FROM WORK Resolve all selects in UNION. When remove_from is true, we don't generate a setop plan and therefore all except the first select remain without ordinals resolved. This leads to their output buing null at execution time and the result is null pointer exception. The fix is to let the LogicAppend find the correct query in the setop tree for each plan and save it's selections when LogicAppend is created.
Override ResolveColumnOrdinal in LogicAppend and using the saved
selections, resolve as many ordianls from each child as the number of
ordinals in the original reqOutput.

Set the yop level LogicAppend's outputs to those of the first child.
The code is rather tricky. More detailed explanation may be needed in
the code.