REMOVE_FROM WORK Handle aggregates appearing in group by after transforming FromQuery. If aggregates are seen in the group by, check if each of them is a select expression in one of the FromQuery's in scope and if so, throw no error.
When resolving ordinals in LogicAgg, if there are aggregates in the
group by, remove each aggregate, each child of AggrRef and other
expression and make a new group by list, save the original group by and
null it out. Turn the newGroupBy as required from child. Once the child
resolve is done, set back the saved groupby as group by and reolve it.

This has a side effect on Agg(Agg(Agg(...))). In general these queries
will throw various exceptions but if the arguments of each of the top
level aggragtes are added as group by of the top level select, they work.