Fix null handling in IN  expr and subquery (issue #246)
NULL comparison only works with x IS NULL and x IS NOT NULL. Only in these cases the result is TRUE or FALSE depending on x being NULL or NOT NULL. Every other way of comparing something against NULL results in NULL. This standard SQL and Postgres follows the standard. We should do the same.