Andb parser int (#281)
Simple binding of FROM (single table) is working.

References to non existing tables raise error.

Joins are not supported:
1) Old style joins give the errors "JOIN not supported", and
2) Joins with JOIN keyword give syntax error.

Header files contaiun only simple function defintions, mainly the ones that don't need to know the size of
class other than what is defined within that file. This is to avoid the problem of incomplete types when headers are
included in a specific order and sometimes this ordering gets quite complicated.