ANDB parser integration into native code.
Execution framework is now connected and PhysicScan _works_.
PhysicScan at the moment has no table source so it simply gives a row
each time Exec is called. This is not good for the machine because at
the moment no matter what table select runs on there will be a 5M+ rows
collected under a minute. Nothing is printed so there is no indication
of memory filling up.

All deletes are disabled, the idea is to use PMR for all allocations and
destructors be noops.

Some problems found under Linux (Ubuntu 20.10, clang 11 and GNU c++ 10.2)
are merged in.